Add workflow to automatically build and push the test container
To be able to automate the tests made for #8, this workflow will build and push a Docker container
to be ready to use on any push to main or open pull request. It also adds a tag depending on if it's
a push to main or a PR so the developers can run their tests again a particular build. This closes #25.